Fractional abundance

Generate an isotope abundance object. See isotopia for general information on initializing and converting isotope data objects.

abundance(..., major = get_iso_opts("default_major"), compound = "",
  notation = get_iso_opts("default_ab_notation"), weight = numeric(),
  single_as_df = FALSE)

ab(..., major = get_iso_opts("default_major"), compound = "",
  notation = get_iso_opts("default_ab_notation"), weight = numeric(),
  single_as_df = FALSE)

Arguments

...
- numeric vectors (can be named) to turn into isotope abundance objects

major
- name of the major isotope in the isotope system [optional], only of importance if converting from abundance to ratio or delta value, and want automatic name propagation

compound
- name of the compound the isotopic values belong to [optional]

notation
- what notation the abundance is in ('raw' values or 'percent'), see switch_notation for details

Details

The ab function is a shorthand for abundance but otherwise identical.

See also