Checks for isotope value objects

Checks for different kinds of isotope value objects. All checks recognize both the vector (single isotope value) and the data.frame (isotope system) version of an isotope value object. is.isosys(obj) can be used to make the distinction between the two.

is.iso(obj)

is.isoval(obj)

is.isosys(obj)

is.ratio(obj)

is.abundance(obj)

is.delta(obj)

is.intensity(obj)

is.ff(obj)

is.weighted(iso)

Arguments

obj
- object to test

Details

is.iso checks whether the object is an isotope value object of any kind. Returns TRUE if it is (e.g. ratio, abundance, delta, etc. - single or system of values), FALSE otherwise. is.isoval checks whether the object is a single isotope value. Returns TRUE if it's a single isotope value object (of any kind, ratio, abundance, delta, etc.) and FALSE otherwise. is.isosys checks whether the object is a an isotope system. Returns TRUE if it's an isotope system (of any kind, ratios, abundances, deltas, etc.) and FALSE otherwise. is.ratio checks whether the object is an isotope ratio object. Returns TRUE if it's a single isotope ratio object or an isotope system of ratios, FALSE otherwise. is.abundance checks whether the object is an isotope abundance object. Returns TRUE if it's a single isotope abundance object or an isotope system of abundances, FALSE otherwise. is.delta checks whether the object is a delta value object. Returns TRUE if it's a single delta value object or an isotope system of delta values, FALSE otherwise. is.intensity checks whether the object is an ion intensity object. Returns TRUE if it's a single ion intensity object or an isotope system of ion intensities, FALSE otherwise. is.ff checks whether the object is an fractionation factor value object. Returns TRUE if it's a single fractionation factor value object or an isotope system of fractionation factor values, FALSE otherwise. is.weighted checks if an isotope object is weighted. An object counts as weighted if any of the weights associated with the data values is != 1, that means only objects whose weights are ALL 1 is considered unweighted.

Examples

is.weighted(ratio(0.2)) # returns FALSE

#> [1] FALSE

is.weighted(ratio(0.2, weight = 1)) # returns FALSE

#> [1] FALSE

is.weighted(ratio(c(0.1, 0.2), weight = c(1,2))) # returns TRUE

#> [1] TRUE