Isotopia options

This allows specifying and retrieving default values for newly created isotopia objects.

set_iso_opts(default_ab_notation = c("raw", "percent"),
  default_ff_notation = c("alpha", "eps", "permil", "ppm"),
  default_delta_notation = c("raw", "permil", "ppm"),
  default_intensity_unit = "", default_major = "", default_minor = "",
  exact_mass_balance = FALSE, standards = c())

get_iso_opts(opts)

Arguments

default_ab_notation
default notation of abundance objects, see switch_notation for details

default_ff_notation
default notation of fractionation factors

default_delta_notation
default notation of delta values

default_intensity_unit
default unit for intensity values

default_major
default major isotope on all isotope objects

exact_mass_balance
NOT IMPLEMENTED YET! If enabled, mass balance calculations with delta values (i.e. mass_balance(delta, delta, delta... or delta() + delta()) will always be performed exact by converting to natural abundances first and making the addition in abundance space (will be converted back to delta value afterwards). This is only possible if the ref_ratio in the delta values is set and will lead to an error if attempted without the reference ratios set.

If disabled, mass balance calculations with delta values (delta() + delta()) will be performed in delta space (which is not exact but the discrepancy is negligible unless the minor isotopes in an isotope system make up a significant portion)

see register_standard for details

standards
isotope ratio objects to register as standards

default
minor default minor isotope on all isotope objects

Details

get_iso_opts allows retrieval of all or individual isotopia options. Returns a single value if only one option is requested, a named list if multiple

Note

Default options are the following and are set during package loading together with the default standards set_iso_opts( default_ab_notation = "raw", default_ff_notation = "alpha", default_delta_notation = "permil", default_intensity_unit = "", default_major = "", default_minor = "", exact_mass_balance = FALSE )

See also

Other options: register_standard

Examples

get_iso_opts("standards") # get a table of all standards
get_iso_opts(c("default_major", "default_minor")) # get a named list with the 
default major and minor isotopes

#> Error: <text>:3:9: unexpected symbol
#> 2: get_iso_opts(c("default_major", "default_minor")) # get a named list with the 
#> 3: default major
#>            ^