Information about an isotopic data object

Get information about the name, label and units of an isotopic data object.

get_name(object)

# S4 method for ANY
get_name(object)

# S4 method for Isoval
get_name(object)

# S4 method for Ratio
get_name(object)

# S4 method for Abundance
get_name(object)

# S4 method for FractionationFactor
get_name(object)

# S4 method for Delta
get_name(object)

get_units(object)

# S4 method for Isoval
get_units(object)

# S4 method for Intensity
get_units(object)

get_label(object)

# S4 method for ANY
get_label(object)

# S4 method for Isoval
get_label(object)

# S4 method for FractionationFactor
get_label(object)

# S4 method for Delta
get_label(object)

# S4 method for Isosys
get_label(object)

Details

get_name() returns the name of an isotopic data object get_units() provides the units of an isotope data object depending on the object type and notation get_label() provides the full label of an isotope data object

See also

Other data type attributes: get_value, get_weighted_value, get_weight, set_attrib, switch_notation, weight

Examples

# NOT RUN {
get_label(ratio(...))
get_label(abundance(...))
get_label(isosys(ratio(...), ratio(...))
# }