Isotope standards

Isotopia provides functionality to register and retrieve isotope standards. Registered standards can be used for automatic conversions of, for example, delta values which have attributes that match a standard.

register_standard(ratio)

get_standards(minor = NULL, major = NULL, name = NULL)

get_standard(minor = NULL, major = NULL, name = NULL)

Arguments

ratio
- a ratio object with 'minor', and 'major' isotope as well as 'compound' (the name of the standard) attributes all defined

minor
- character vector of minor isotope names to search for

major
- character vector of major isotope names to search for

name
- character vector of standards names to search for

Value

list of ratio objects

Details

Use register_standard() to register an isotope standard. This can be useful for keeping track of standards you use internally and will also allow conversions from delta to e.g. ratio to automatically try to find the approriate standard for the conversion from the registered values.

Use get_stanards to retrieve any number of registered isotope standards that can be identified with the provided search terms. For an overview table of all standards (rather than the actual ratio objects), use get_iso_opts("standards") instead.

get_standard is the same as get_standards except that it returns a single object from the found standards and throws an error if the search criteria did not yield exactly one.

See also