Switch notation

Convert from one notation to another for an isotope data object.

switch_notation(iso, to)

# S4 method for ANY,ANY,ANY
switch_notation(iso, to = NULL, from = NULL)

# S4 method for numeric,Notation_raw,Notation_raw
switch_notation(iso, to = NULL,
  from = NULL)

# S4 method for numeric,Notation_percent,Notation_percent
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_alpha,Notation_alpha
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_eps,Notation_eps
switch_notation(iso, to = NULL,
  from = NULL)

# S4 method for numeric,Notation_permil,Notation_permil
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_ppm,Notation_ppm
switch_notation(iso, to = NULL,
  from = NULL)

# S4 method for numeric,Notation_percent,Notation_raw
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_raw,Notation_percent
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_eps,Notation_raw
switch_notation(iso, to = NULL,
  from = NULL)

# S4 method for numeric,Notation_raw,Notation_eps
switch_notation(iso, to = NULL,
  from = NULL)

# S4 method for numeric,Notation_eps,Notation_alpha
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_alpha,Notation_eps
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_alpha,Notation_raw
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_permil,Notation_raw
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_raw,Notation_permil
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_permil,Notation_eps
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_eps,Notation_permil
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_permil,Notation_alpha
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_alpha,Notation_permil
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for numeric,Notation_ppm,ANY
switch_notation(iso, to = NULL,
  from = NULL)

# S4 method for numeric,ANY,Notation_ppm
switch_notation(iso, to = NULL,
  from = NULL)

# S4 method for Isoval,ANY,missing
switch_notation(iso, to = NULL, from = NULL)

# S4 method for Isoval,Notation_raw,missing
switch_notation(iso, to = NULL,
  from = NULL)

# S4 method for Abundance,Notation_percent,missing
switch_notation(iso, to = NULL,
  from = NULL)

# S4 method for Delta,Notation_permil,missing
switch_notation(iso, to = NULL,
  from = NULL)

# S4 method for Delta,Notation_ppm,missing
switch_notation(iso, to = NULL,
  from = NULL)

# S4 method for FractionationFactor,Notation_alpha,missing
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for FractionationFactor,Notation_eps,missing
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for FractionationFactor,Notation_raw,missing
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for FractionationFactor,Notation_permil,missing
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for FractionationFactor,Notation_ppm,missing
switch_notation(iso,
  to = NULL, from = NULL)

# S4 method for Isoval,character,missing
switch_notation(iso, to = NULL,
  from = NULL)

# S4 method for Isosys,character,missing
switch_notation(iso, to = NULL,
  from = NULL)

Arguments

iso
isotopic data object (ff, abundance, delta)

to
which notation to convert to

Value

isotope object with converted notation, an error if it is not a valid conversion

Details

Valid notations depend on the data type:

abundance: 'raw', 'percent'

delta: 'raw', 'permil', 'ppm'

fractionation_factor: 'alpha', 'eps', 'permil', 'ppm'

See also

Other data type attributes: get_name, get_value, get_weighted_value, get_weight, set_attrib, weight