Fractionation factor

Calculate/convert to an isotope fractionation_factor

to_ff(iso1, iso2)

# S4 method for ANY,ANY
to_ff(iso1, iso2)

# S4 method for FractionationFactor,missing
to_ff(iso1, iso2)

# S4 method for FractionationFactors,missing
to_ff(iso1, iso2)

# S4 method for Ratio,Ratio
to_ff(iso1, iso2)

# S4 method for Delta,missing
to_ff(iso1, iso2)

# S4 method for Delta,Delta
to_ff(iso1, iso2)

Arguments

iso1
the top compound in the fractionation factor

iso2
the bottom compound in the fractionation factor

Value

isotope fraction_factor object if parameters can be converted, an error otherwise

Details

The to_ff(...) function calculates the fractionation factor between two isotope data objects (for example two delta values, two ratio, or two ff). All calculatinos are only permissible if the isotope values have matching attributes and fractionation factors will be returend in the default notation (see set_iso_opts for details)

Note

Some of the conversions are also implemented in arithmetic shorthand, for example to generate an fractionation factor in alpha notation from two ratios to_ff(ratio(), ratio()) is the same as ratio() / ratio(). See arithmetic for details.

See also

Other data type conversions: to_abundance, to_delta, to_ratio